+++ packages/commerce-sdk-react/src/hooks/ShopperContexts/cache.ts
@@ -6,8 +6,10 @@
  */
 import {ApiClients, CacheUpdateMatrix} from '../types'
 import {getShopperContext} from './queryKeyHelpers'
+import {CLIENT_KEYS} from '../../constant'
 
-type Client = ApiClients['shopperContexts']
+const CLIENT_KEY = CLIENT_KEYS.SHOPPER_CONTEXTS
+type Client = NonNullable<ApiClients[typeof CLIENT_KEY]>
 
 // TODO: Complete cache invalidation https://gus.lightning.force.com/lightning/_classic/%2Fa07EE00001NoYplYAF
 export const cacheUpdateMatrix: CacheUpdateMatrix<Client> = {