+++ packages/commerce-sdk-react/src/hooks/ShopperContexts/mutation.test.ts
@@ -18,6 +18,7 @@ import {
     waitAndExpectError,
     waitAndExpectSuccess
 } from '../../test-utils'
+import {CLIENT_KEYS} from '../../constant'
 
 jest.mock('../../auth/index.ts', () => {
     return jest.fn().mockImplementation(() => ({
@@ -26,7 +27,8 @@ jest.mock('../../auth/index.ts', () => {
     }))
 })
 
-type Client = ApiClients['shopperContexts']
+const CLIENT_KEY = CLIENT_KEYS.SHOPPER_CONTEXTS
+type Client = NonNullable<ApiClients[typeof CLIENT_KEY]>
 
 const contextEndpoint = '/shopper/shopper-context/'